LayoutTests/inspector/unit-tests/globals-unhandled-rejection-in-test-suite-expected.txt

Test that the unhandled promise rejection hook will immediately terminate the test and print the associated exception and stack trace.


== Running test suite: ReportUnhandledRejection
-- Running test case: CatchRejectedPromise
Unhandled promise rejection in inspector page: This promise is rejected with an Error object.

Stack Trace: (suppressed)
